+++ tests/orquestra/vqa/estimation/cvar_test.py
@@ -6,7 +6,7 @@ from orquestra.quantum.api.estimation import EstimationTask
 from orquestra.quantum.api.estimator_contract import ESTIMATOR_CONTRACTS
 from orquestra.quantum.circuits import Circuit, H, X
 from orquestra.quantum.operators import PauliTerm
-from orquestra.quantum.symbolic_simulator import SymbolicSimulator
+from orquestra.quantum.runners import SymbolicSimulator
 from orquestra.quantum.testing import MockCircuitRunner
 
 from orquestra.vqa.estimation.cvar import CvarEstimator
